* commit 'b0501ab66ddf82f5d3b6aa99734e14a1e27c916a':
  Fix host 64-bit build.

+++ libziparchive/zip_archive_test.cc
@@ -209,7 +209,8 @@ TEST(ziparchive, ExtractToFile) {
                       sizeof(kATxtContents)));
 
   // Assert that the total length of the file is sane
-  ASSERT_EQ(data_size + sizeof(kATxtContents), lseek64(fd, 0, SEEK_END));
+  ASSERT_EQ(data_size + static_cast<ssize_t>(sizeof(kATxtContents)),
+            lseek64(fd, 0, SEEK_END));
 
   close(fd);
 }
@@ -247,4 +248,3 @@ int main(int argc, char** argv) {
 
   return RUN_ALL_TESTS();
 }
-
